LayoutTests/fast/masking/clip-path-selection-expected.txt

PASS selection.toString() is ""
PASS selection.toString() is 'was'
PASS selection.toString() is ""
PASS selection.toString() is 'was'
PASS selection.toString() is ""
PASS selection.toString() is 'was'
PASS selection.toString() is ""
PASS selection.toString() is 'was'
PASS successfullyParsed is true

TEST COMPLETE
